Add support for Arduino zero
WARNING ZERO 1: Debug must be on for this to work.

WARNING ZERO 2: USB host interface can block up to 10 seconds waiting for
input.

WARNING ZERO 3: Do not update the USB host library. This will install the
USB host library for Due which does not work on Zero. Use the Zero built-in
USB host library.

The Zero branch works for Due as well. USB host library for Due does not
have problems 1, 2, and 3.